Warning Fixes
Fixing the deprecated-enum-enum-conversion warnings.

+++ Net/src/WebSocket.cpp
@@ -148,7 +148,7 @@ void WebSocket::shutdown(Poco::UInt16 statusCode, const std::string& statusMessa
 	Poco::BinaryWriter writer(ostr, Poco::BinaryWriter::NETWORK_BYTE_ORDER);
 	writer << statusCode;
 	writer.writeRaw(statusMessage);
-	sendFrame(buffer.begin(), static_cast<int>(ostr.charsWritten()), FRAME_FLAG_FIN | FRAME_OP_CLOSE);
+	sendFrame(buffer.begin(), static_cast<int>(ostr.charsWritten()), static_cast<int>(FRAME_FLAG_FIN) | static_cast<int>(FRAME_OP_CLOSE));
 }
